+++ app/Http/Controllers/SessionController.php
@@ -2,17 +2,15 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\RoomRequest;
 use App\Http\Requests\SessionRequest;
 use App\Models\Movie;
 use App\Models\Room;
 use App\Models\Session;
-use Carbon\Carbon;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class SessionController extends Controller
 {
@@ -28,10 +26,11 @@ class SessionController extends Controller
         $this->middleware('auth')->only('edit');
 
     }
+
     public function index(): Factory|View|Application
     {
 
-        return view('sessions.index', ['sessions'=>Session::orderBy('start_time', 'ASC')->get(), "rooms"=>Room::all(), "movies"=>Movie::all()]);
+        return view('sessions.index', ['sessions' => Session::orderBy('start_time', 'ASC')->get(), "rooms" => Room::all(), "movies" => Movie::all()]);
     }
 
     /**
@@ -42,7 +41,7 @@ class SessionController extends Controller
      */
     public function create(Session $session): View|Factory|Application
     {
-        return view('sessions.create', ['session'=>$session, "rooms"=>Room::orderBy('name', 'ASC')->get(), "movies"=>Movie::orderBy('title', 'ASC')->get()]);
+        return view('sessions.create', ['session' => $session, "rooms" => Room::orderBy('name', 'ASC')->get(), "movies" => Movie::orderBy('title', 'ASC')->get()]);
 
     }
 
@@ -54,7 +53,8 @@ class SessionController extends Controller
      */
     public function store(SessionRequest $request): RedirectResponse
     {
-        $data=$request->all();
+        $data = $request->all();
+        $data["user_id"] = Auth::user()->id;
         Session::create($data);
         return redirect()->route("session.index")
             ->with("ok", __("Session has been saved"));
@@ -79,7 +79,7 @@ class SessionController extends Controller
      */
     public function edit(Session $session): View|Factory|Application
     {
-        return view('sessions.edit', ['session'=>$session, "rooms"=>Room::orderBy('name', 'ASC')->get(), "movies"=>Movie::orderBy('title', 'ASC')->get()]);
+        return view('sessions.edit', ['session' => $session, "rooms" => Room::orderBy('name', 'ASC')->get(), "movies" => Movie::orderBy('title', 'ASC')->get()]);
 
 
     }
@@ -93,21 +93,32 @@ class SessionController extends Controller
      */
     public function update(SessionRequest $request, Session $session): RedirectResponse
     {
-        $data = $request->all();
-        $session->update($data);
-        return redirect()->route("session.index")
-            ->with("ok", __("Session has been updated"));
+        if ($session->user_id == Auth::user()->id) {
+
+            $data = $request->all();
+            $session->update($data);
+            return redirect()->route("session.index")
+                ->with("ok", __("Session has been updated"));
+        }
+
+        abort(403, 'Unauthorized action.');
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param Session $session
-     * @return RedirectResponse
+     * @return mixed
      */
-    public function destroy(Session $session): RedirectResponse
+    public function destroy(Session $session): mixed
     {
-        $session->delete();
-        return redirect()->back();
+        if ($session->user_id == Auth::user()->id) {
+
+
+            $session->delete();
+            return redirect()->json();
+        }
+
+        abort(403, 'Unauthorized action.');
     }
 }
